b. initialState The initialState property defines the initial state of the slice. It is an object that includes the following properties:
open: A boolean that represents whether a certain part of the app is open or not. It is initialized to false.
selectedMail: Initially null, this property holds the currently selected email.
searchText: A string used to store the current search text, initialized to an empty string.
emails: This property is initialized to null but could be an array or object depending on the application's logic. It is intended to hold the list of emails.
authUser: Initially null, this property holds the authenticated user's information.