c. reducers The reducers property is an object where each key is a reducer function that can modify the state. Reducers are responsible for handling actions and updating the state accordingly. Each reducer function receives the current state and an action as arguments.
Reducer Functions:
setOpen: This reducer updates the open property in the state. It sets the open property to the value provided in the action.payload.
setSelectedMail: This reducer updates the selectedMail property with the value provided in the action.payload.
setSearchText: This reducer sets the searchText property to the value provided in the action.payload.
setEmails: This reducer updates the emails property with the value provided in the action.payload.
setAuthUser: This reducer sets the authUser property to the value provided in the action.payload.