Exporting the Reducer Finally, the reducer generated by createSlice is exported as the default export of the module. The reducer is responsible for handling the actions and updating the state based on the reducer functions defined in the reducers property.
Summary
Importing createSlice: Brings in the utility to create a slice of the Redux store.
Defining the Slice: Specifies the slice name ("app"), the initial state, and the reducers.
Initial State: Defines the initial values for open, selectedMail, searchText, emails, and authUser.
Reducers: Functions that handle actions and update the state (setOpen, setSelectedMail, setSearchText, setEmails, setAuthUser).
Exporting Actions: Makes the action creators available for use in the application.
Exporting the Reducer: Provides the reducer to be included in the Redux store configuration.